move response formatting to execution service
response formatting in supergraph service means that the execution
service is produing invalid responses, which we should avoid
